Working on some additional functionality and refactoring

+++ CarSharingApp.PublicApi/Controllers/CustomersController.cs
@@ -1,14 +1,14 @@
 ﻿using CarSharingApp.Application.Contracts.Customer;
 using CarSharingApp.Application.Interfaces;
 using CarSharingApp.Domain.Entities;
+using CarSharingApp.Infrastructure.Authentication;
 using ErrorOr;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CarSharingApp.PublicApi.Controllers
 {
-    [Authorize]
-    public class CustomersController : ApiController
+    public sealed class CustomersController : ApiController
     {
         private readonly ICustomerService _customerService;
 
@@ -18,6 +18,7 @@ namespace CarSharingApp.PublicApi.Controllers
         }
 
         [HttpPost]
+        [AllowAnonymous]
         public async Task<IActionResult> CreateCustomer(CreateCustomerRequest request)
         {
             ErrorOr<Customer> requestToCustomerResult = _customerService.From(request);
@@ -40,20 +41,40 @@ namespace CarSharingApp.PublicApi.Controllers
         }
 
         [HttpGet("{id:guid}")]
+        [Authorize(Roles = "Administrator, Customer")]
         public async Task<IActionResult> GetCustomer(Guid id)
         {
+            if (!IsRequestAllowed(id)) 
+            {
+                return Forbid();
+            }
+
             ErrorOr<Customer> getCustomerResult = await _customerService.GetCustomerAsync(id);
 
             return getCustomerResult.Match(
                 customer => Ok(MapCustomerResponse(customer)),
                 errors => Problem(errors));
-
         }
         
-        [HttpPut("{id:guid}")]
-        public async Task<IActionResult> UpdateCustomer(Guid id, UpdateCustomerRequest request)
+        [HttpPut("Info/{id:guid}")]
+        [Authorize(Roles = "Administrator, Customer")]
+        public async Task<IActionResult> UpdateCustomerInfo(Guid id, UpdateCustomerInfoRequest request)
         {
-            ErrorOr<Customer> requestToCustomerResult = Error.Failure(); //_customerService.From(id, Credentials, request);
+            if (!IsRequestAllowed(id))
+            {
+                return Forbid();
+            }
+
+            ErrorOr<Customer> getCustomerResult = await _customerService.GetCustomerAsync(id);
+
+            if (getCustomerResult.IsError)
+            {
+                return Problem(getCustomerResult.Errors);
+            }
+
+            Customer notUpdatedCustomerYet = getCustomerResult.Value;
+
+            ErrorOr<Customer> requestToCustomerResult = _customerService.From(notUpdatedCustomerYet, request);
 
             if (requestToCustomerResult.IsError)
             {
@@ -62,7 +83,80 @@ namespace CarSharingApp.PublicApi.Controllers
 
             Customer customer = requestToCustomerResult.Value;
 
-            ErrorOr<Updated> updateCustomerResult = await _customerService.UpdateCustomerAsync(customer);
+            ErrorOr<Updated> updateCustomerInfoResult = await _customerService.UpdateCustomerInfoAsync(customer);
+
+            return updateCustomerInfoResult.Match(
+                updated => NoContent(),
+                errors => Problem(errors));
+        }
+
+        [HttpPut("Credentials/{id:guid}")]
+        [Authorize(Roles = "Administrator, Customer")]
+        public async Task<IActionResult> UpdateCustomerCredentials(Guid id, UpdateCustomerCredentialsRequest request)
+        {
+            if (!IsRequestAllowed(id))
+            {
+                return Forbid();
+            }
+
+            ErrorOr<Customer> getCustomerResult = await _customerService.GetCustomerAsync(id);
+
+            if (getCustomerResult.IsError)
+            {
+                return Problem(getCustomerResult.Errors);
+            }
+
+            Customer notUpdatedCustomerYet = getCustomerResult.Value;
+
+            ErrorOr<Customer> requestToCustomerResult = _customerService.From(notUpdatedCustomerYet, request);
+
+            if (requestToCustomerResult.IsError)
+            {
+                return Problem(requestToCustomerResult.Errors);
+            }
+
+            Customer customer = requestToCustomerResult.Value;
+
+            ErrorOr<Updated> updateCustomerResult = await _customerService.UpdateCustomerPasswordAsync(customer);
+
+            return updateCustomerResult.Match(
+                updated => NoContent(),
+                errors => Problem(errors));
+        }
+
+        [HttpPut("Password/{id:guid}")]
+        [Authorize(Roles = "Administrator, Customer")]
+        public async Task<IActionResult> UpdateCustomerPassword(Guid id, UpdateCustomerPasswordRequest request)
+        {
+            if (!IsRequestAllowed(id))
+            {
+                return Forbid();
+            }
+
+            ErrorOr<Customer> getCustomerResult = await _customerService.GetCustomerAsync(id);
+
+            if (getCustomerResult.IsError)
+            {
+                return Problem(getCustomerResult.Errors);
+            }
+
+            Customer customerWithOldPassword = getCustomerResult.Value;
+
+            ErrorOr<Customer> requestToCustomerResult = _customerService.From(customerWithOldPassword, request);
+            if (requestToCustomerResult.IsError)
+            {
+                return Problem(requestToCustomerResult.Errors);
+            }
+
+            Customer customerWithUpdatedPassword = requestToCustomerResult.Value;
+
+            ErrorOr<Success> checkoutIfOldPasswordMatchResult = await _customerService.CompareCustomerOldPasswordWithExistingOne(id, request.currentPassword);
+            if (checkoutIfOldPasswordMatchResult.IsError)
+            {
+                return Problem(checkoutIfOldPasswordMatchResult.Errors);
+            }
+
+            ErrorOr<Updated> updateCustomerResult = await _customerService.UpdateCustomerPasswordAsync(customerWithUpdatedPassword);
 
             return updateCustomerResult.Match(
                 updated => NoContent(),
@@ -70,8 +164,23 @@ namespace CarSharingApp.PublicApi.Controllers
         }
 
         [HttpDelete("{id:guid}")]
+        [Authorize(Roles = "Administrator, Customer")]
         public async Task<IActionResult> DeleteCustomer(Guid id)
         {
+            if (!IsRequestAllowed(id))
+            {
+                return Forbid();
+            }
+
+            ErrorOr<Customer> getCustomerResult = await _customerService.GetCustomerAsync(id);
+
+            if (getCustomerResult.IsError)
+            {
+                return Problem(getCustomerResult.Errors);
+            }
+
+            Customer notDeletedCustomerYet = getCustomerResult.Value;
+
             ErrorOr<Deleted> deleteCustomerResult = await _customerService.DeleteCustomerAsync(id);
 
             return deleteCustomerResult.Match(
@@ -85,19 +194,25 @@ namespace CarSharingApp.PublicApi.Controllers
                 customer.Id,
                 customer.FirstName,
                 customer.LastName,
-                customer.Country,
-                customer.City,
                 customer.Address,
                 customer.PhoneNumber,
                 customer.DriverLicenseIdentifier,
-                customer.ProfileDescription,
-                customer.ProfileImage,
-                customer.Postcode,
-                customer.VehiclesOrdered,
-                customer.VehiclesShared,
+                customer.Profile,
+                customer.Statistics,
                 customer.HasAcceptedNewsSharing,
-                customer.IsOnline,
                 customer.Credentials);
+        }
+
+        private bool IsRequestAllowed(Guid requestedId)
+        {
+            JwtClaims? jwtClaims = GetJwtClaims();
+
+            if (jwtClaims == null || (jwtClaims.Id != requestedId.ToString() && jwtClaims.Role != "Administrator"))
+            {
+                return false;
+            }
+
+            return true;
         }
     }
 }
